feat: support event count limits and link count limits for spans
Without this, we can easily grow memory usage in SpanBuilder infinitely.
This is likely the cause of a number of excessive memory usage reports.

Ideally, we'd pick up the limits here from the opentelemetry
configuration on the Tracer, but the opentelemetry::tracer::Tracer trait
doesn't appear to provide a way to obtain the configuration so there
will be some potential for unexpected conflicts between the two
configurations.